+++ src/main/scala/com/madewithtea/mockedstreams/MockedStreams.scala
@@ -14,7 +14,6 @@
   * See the License for the specific language governing permissions and
   * limitations under the License.
   */
-
 package com.madewithtea.mockedstreams
 
 import java.time.Instant
@@ -22,10 +21,14 @@ import java.util.{Properties, UUID}
 
 import org.apache.kafka.clients.consumer.ConsumerRecord
 import org.apache.kafka.common.serialization.Serde
-import org.apache.kafka.streams.state.ReadOnlyWindowStore
-import org.apache.kafka.streams.test.ConsumerRecordFactory
-import org.apache.kafka.streams.{StreamsConfig, Topology, TopologyTestDriver => Driver}
 import org.apache.kafka.streams.scala.StreamsBuilder
+import org.apache.kafka.streams.state.ValueAndTimestamp
+import org.apache.kafka.streams.test.ConsumerRecordFactory
+import org.apache.kafka.streams.{
+  StreamsConfig,
+  Topology,
+  TopologyTestDriver => Driver
+}
 
 import scala.collection.JavaConverters._
 import scala.collection.immutable
@@ -34,12 +37,15 @@ object MockedStreams {
 
   def apply() = Builder()
 
-  case class Builder(topology: Option[() => Topology] = None,
-                     configuration: Properties = new Properties(),
-                     stateStores: Seq[String] = Seq(),
-                     inputs: List[ConsumerRecord[Array[Byte], Array[Byte]]] = List.empty) {
+  case class Builder(
+      topology: Option[() => Topology] = None,
+      configuration: Properties = new Properties(),
+      stateStores: Seq[String] = Seq(),
+      inputs: List[ConsumerRecord[Array[Byte], Array[Byte]]] = List.empty
+  ) {
 
-    def config(configuration: Properties): Builder = this.copy(configuration = configuration)
+    def config(configuration: Properties): Builder =
+      this.copy(configuration = configuration)
 
     def topology(func: StreamsBuilder => Unit): Builder = {
       val buildTopology = () => {
@@ -54,71 +60,113 @@ object MockedStreams {
 
     def stores(stores: Seq[String]): Builder = this.copy(stateStores = stores)
 
-    def input[K, V](topic: String, key: Serde[K], value: Serde[V], records: Seq[(K, V)]): Builder =
+    def input[K, V](
+        topic: String,
+        key: Serde[K],
+        value: Serde[V],
+        records: Seq[(K, V)]
+    ): Builder =
       _input(topic, key, value, Left(records))
 
-    def inputWithTime[K, V](topic: String, key: Serde[K], value: Serde[V], records: Seq[(K, V, Long)]): Builder =
+    def inputWithTime[K, V](
+        topic: String,
+        key: Serde[K],
+        value: Serde[V],
+        records: Seq[(K, V, Long)]
+    ): Builder =
       _input(topic, key, value, Right(records))
 
-    def output[K, V](topic: String, key: Serde[K], value: Serde[V], size: Int): immutable.IndexedSeq[(K, V)] = {
+    def output[K, V](
+        topic: String,
+        key: Serde[K],
+        value: Serde[V],
+        size: Int
+    ): immutable.IndexedSeq[(K, V)] = {
       if (size <= 0) throw new ExpectedOutputIsEmpty
       withProcessedDriver { driver =>
         (0 until size).flatMap { _ =>
           Option(driver.readOutput(topic, key.deserializer, value.deserializer)) match {
             case Some(record) => Some((record.key, record.value))
-            case None => None
+            case None         => None
           }
         }
       }
     }
 
-    def outputTable[K, V](topic: String, key: Serde[K], value: Serde[V], size: Int): Map[K, V] =
+    def outputTable[K, V](
+        topic: String,
+        key: Serde[K],
+        value: Serde[V],
+        size: Int
+    ): Map[K, V] =
       output[K, V](topic, key, value, size).toMap
 
-    def stateTable(name: String): Map[Nothing, Nothing] = withProcessedDriver { driver =>
-      val records = driver.getKeyValueStore(name).all()
-      val list = records.asScala.toList.map { record => (record.key, record.value) }
-      records.close()
-      list.toMap
+    def stateTable(name: String): Map[Nothing, Nothing] = withProcessedDriver {
+      driver =>
+        val records = driver.getKeyValueStore(name).all()
+        val list = records.asScala.toList.map { record =>
+          (record.key, record.value)
+        }
+        records.close()
+        list.toMap
     }
 
     /**
       * @throws IllegalArgumentException if duration is negative or can't be represented as long milliseconds
       */
-    def windowStateTable[K, V](name: String,
-                               key: K,
-                               timeFrom: Long = 0,
-                               timeTo: Long = Long.MaxValue): Map[java.lang.Long, V] = {
-      windowStateTable[K, V](name, key, Instant.ofEpochMilli(timeFrom), Instant.ofEpochMilli(timeTo))
+    def windowStateTable[K, V](
+        name: String,
+        key: K,
+        timeFrom: Long = 0,
+        timeTo: Long = Long.MaxValue
+    ): Map[java.lang.Long, ValueAndTimestamp[V]] = {
+      windowStateTable[K, V](
+        name,
+        key,
+        Instant.ofEpochMilli(timeFrom),
+        Instant.ofEpochMilli(timeTo)
+      )
     }
 
     /**
       * @throws IllegalArgumentException if duration is negative or can't be represented as long milliseconds
       */
-    def windowStateTable[K, V](name: String,
-                               key: K,
-                               timeFrom: Instant,
-                               timeTo: Instant): Map[java.lang.Long, V] = withProcessedDriver { driver =>
-      val store = driver.getStateStore(name).asInstanceOf[ReadOnlyWindowStore[K, V]]
-      val records = store.fetch(key, timeFrom, timeTo)
-      val list = records.asScala.toList.map { record => (record.key, record.value) }
-      records.close()
-      list.toMap
-    }
+    def windowStateTable[K, V](
+        name: String,
+        key: K,
+        timeFrom: Instant,
+        timeTo: Instant
+    ): Map[java.lang.Long, ValueAndTimestamp[V]] =
+      withProcessedDriver { driver =>
+        val store = driver.getTimestampedWindowStore[K, V](name)
+        val records = store.fetch(key, timeFrom, timeTo)
+        val list = records.asScala.toList.map { record =>
+          (record.key, record.value)
+        }
+        records.close()
+        list.toMap
+      }
 
-    private def _input[K, V](topic: String, key: Serde[K], value: Serde[V],
-                             records: Either[Seq[(K, V)], Seq[(K, V, Long)]]) = {
+    private def _input[K, V](
+        topic: String,
+        key: Serde[K],
+        value: Serde[V],
+        records: Either[Seq[(K, V)], Seq[(K, V, Long)]]
+    ) = {
       val keySer = key.serializer
       val valSer = value.serializer
       val factory = new ConsumerRecordFactory[K, V](keySer, valSer)
 
       val updatedRecords = records match {
-        case Left(withoutTime) => withoutTime.foldLeft(inputs) {
-          case (events, (k, v)) => events :+ factory.create(topic, k, v)
-        }
-        case Right(withTime) => withTime.foldLeft(inputs) {
-          case (events, (k, v, timestamp)) => events :+ factory.create(topic, k, v, timestamp)
-        }
+        case Left(withoutTime) =>
+          withoutTime.foldLeft(inputs) {
+            case (events, (k, v)) => events :+ factory.create(topic, k, v)
+          }
+        case Right(withTime) =>
+          withTime.foldLeft(inputs) {
+            case (events, (k, v, timestamp)) =>
+              events :+ factory.create(topic, k, v, timestamp)
+          }
       }
       this.copy(inputs = updatedRecords)
     }
@@ -126,9 +174,12 @@ object MockedStreams {
     // state store is temporarily created in ProcessorTopologyTestDriver
     private def stream = {
       val props = new Properties
-      props.put(StreamsConfig.APPLICATION_ID_CONFIG, s"mocked-${UUID.randomUUID().toString}")
+      props.put(
+        StreamsConfig.APPLICATION_ID_CONFIG,
+        s"mocked-${UUID.randomUUID().toString}"
+      )
       props.put(StreamsConfig.BOOTSTRAP_SERVERS_CONFIG, "localhost:9092")
-      props.putAll(configuration)
+      configuration.asScala.foreach { case (k, v) => props.put(k, v) }
       new Driver(topology.getOrElse(throw new NoTopologySpecified)(), props)
     }
 
@@ -146,10 +197,15 @@ object MockedStreams {
     }
   }
 
-  class NoTopologySpecified extends Exception("No topology specified. Call topology() on builder.")
+  class NoTopologySpecified
+      extends Exception("No topology specified. Call topology() on builder.")
 
-  class NoInputSpecified extends Exception("No input fixtures specified. Call input() method on builder.")
+  class NoInputSpecified
+      extends Exception(
+        "No input fixtures specified. Call input() method on builder."
+      )
 
-  class ExpectedOutputIsEmpty extends Exception("Output size needs to be greater than 0.")
+  class ExpectedOutputIsEmpty
+      extends Exception("Output size needs to be greater than 0.")
 
 }